+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+- Fixed bug to ignore additional characters on valid tax code.
 
 ## [2.9.1] - 2023-03-13Z
 - Downgrade Kotlin version for compatibility.